Work on demonstrator setup and back-end

+++ sys-buyer/src/main/java/wm_demo/buyer/DataOfferNew.java
@@ -10,7 +10,7 @@ import static se.arkalix.dto.DtoEncoding.JSON;
 @DtoReadableAs(JSON)
 @DtoWritableAs(JSON)
 public interface DataOfferNew {
-    Optional<Integer> id();
+    Optional<Long> id();
 
     boolean drilled();
